+++ include/gpu/qualityscoreweights.hpp
@@ -7,7 +7,7 @@ namespace care {
 namespace gpu {
 
 #ifdef __NVCC__
-extern __device__ double d_qscore_to_weight[256];
+extern __device__ float d_qscore_to_weight[256];
 #endif
 
 void init_weights(const std::vector<int>& deviceIds);
